@@ -1,6 +1,6 @@
 {
   "name": "samlify",
-  "version": "2.6.1",
+  "version": "2.8.10",
   "description": "High-level API for Single Sign On (SAML 2.0)",
   "main": "build/index.js",
   "keywords": [
@@ -12,7 +12,7 @@
   ],
   "typings": "types/index.d.ts",
   "scripts": {
-    "build": "make rebuild",
+    "build": "yarn audit;make rebuild",
     "docs": "docsify serve -o docs",
     "lint": "tslint -p .",
     "lint:fix": "tslint -p . --fix",
@@ -31,42 +31,42 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@authenio/xml-encryption": "^0.11.3",
-    "camelcase": "^5.3.1",
-    "deflate-js": "^0.2.3",
-    "node-forge": "^0.8.5",
-    "node-rsa": "^1.0.5",
-    "uuid": "^3.3.2",
+    "@authenio/xml-encryption": "^2.0.2",
+    "@xmldom/xmldom": "^0.8.6",
+    "camelcase": "^6.2.0",
+    "node-forge": "^1.3.0",
+    "node-rsa": "^1.1.1",
+    "pako": "^1.0.10",
+    "uuid": "^8.3.2",
     "xml": "^1.0.1",
-    "xml-crypto": "^1.4.0",
-    "xmldom": "^0.1.27",
-    "xpath": "^0.0.27"
-  },
-  "peerDependencies": {
-    "@authenio/samlify-libxml-xsd": "^1.0.1",
-    "@authenio/samlify-node-xmllint": "^1.0.1",
-    "@authenio/samlify-validate-with-xmllint": "^1.0.1",
-    "@authenio/samlify-xsd-schema-validator": "^1.0.2"
+    "xml-crypto": "^3.0.1",
+    "xpath": "^0.0.32"
   },
   "devDependencies": {
+    "@ava/typescript": "^1.1.1",
     "@types/node": "^11.11.3",
-    "@types/node-forge": "^0.7.4",
-    "@types/uuid": "3.0.0",
-    "@types/xmldom": "^0.1.28",
-    "ava": "^1.2.1",
-    "coveralls": "^3.0.2",
-    "nyc": "^11.9.0",
-    "ts-node": "^8.3.0",
-    "tslint": "^5.12.1",
-    "typescript": "^3.4.5"
+    "@types/node-forge": "^1.0.1",
+    "@types/node-rsa": "^1.1.1",
+    "@types/pako": "^1.0.1",
+    "@types/uuid": "^8.3.4",
+    "@types/xmldom": "^0.1.31",
+    "ava": "^4.1.0",
+    "coveralls": "^3.1.1",
+    "nyc": "^15.1.0",
+    "timekeeper": "^2.2.0",
+    "ts-node": "^10.7.0",
+    "tslint": "^6.1.3",
+    "typescript": "^4.4.2"
   },
   "ava": {
-    "compileEnhancements": false,
     "extensions": [
       "ts"
     ],
     "require": [
       "ts-node/register"
+    ],
+    "files": [
+      "!**/*.d.ts"
     ]
   }
 }
