@@ -7,8 +7,8 @@ import Metadata, { MetadataInterface } from './metadata';
 import { MetadataSpConstructor, MetadataSpOptions } from './types';
 import { namespace, elementsOrder as order } from './urn';
 import libsaml from './libsaml';
-import { isNonEmptyArray, isString } from './utility';
-import * as xml from 'xml';
+import { castArrayOpt, isNonEmptyArray, isString } from './utility';
+import xml from 'xml';
 
 export interface SpMetadataInterface extends MetadataInterface {
 
@@ -36,14 +36,14 @@ export default function(meta: MetadataSpConstructor) {
 export class SpMetadata extends Metadata {
 
   /**
-  * @param  {object/string} meta (either xml string or configuation in object)
+  * @param  {object/string} meta (either xml string or configuration in object)
   * @return {object} prototypes including public functions
   */
   constructor(meta: MetadataSpConstructor) {
 
     const isFile = isString(meta) || meta instanceof Buffer;
 
-    // use object configuation instead of importing metadata file directly
+    // use object configuration instead of importing metadata file directly
     if (!isFile) {
 
       const {
@@ -80,16 +80,12 @@ export class SpMetadata extends Metadata {
         console.warn('Construct service provider - missing signatureConfig');
       }
 
-      if (signingCert) {
-        descriptors.KeyDescriptor!.push(libsaml.createKeySection('signing', signingCert).KeyDescriptor);
-      } else {
-        //console.warn('Construct service provider - missing signing certificate');
+      for(const cert of castArrayOpt(signingCert)) {
+        descriptors.KeyDescriptor!.push(libsaml.createKeySection('signing', cert).KeyDescriptor);
       }
 
-      if (encryptCert) {
-        descriptors.KeyDescriptor!.push(libsaml.createKeySection('encryption', encryptCert).KeyDescriptor);
-      } else {
-        //console.warn('Construct service provider - missing encrypt certificate');
+      for(const cert of castArrayOpt(encryptCert)) {
+        descriptors.KeyDescriptor!.push(libsaml.createKeySection('encryption', cert).KeyDescriptor);
       }
 
       if (isNonEmptyArray(nameIDFormat)) {
@@ -100,10 +96,8 @@ export class SpMetadata extends Metadata {
       }
 
       if (isNonEmptyArray(singleLogoutService)) {
-        let indexCount = 0;
         singleLogoutService.forEach(a => {
           const attr: any = {
-            index: String(indexCount++),
             Binding: a.Binding,
             Location: a.Location,
           };
